refactor: separate constants of parseCSV for each command
