+++ src/record/import/parsers/parseCsv/constants.ts
@@ -1,0 +1,4 @@
+// This LINE_BREAK represents the line break inside the cell
+export const LINE_BREAK = "\n";
+export const SEPARATOR = ",";
+export const PRIMARY_MARK = "*";
